pyproject.toml
[project]
name = "akrocansim"
dynamic = ["version", "description"]
authors = [{name = "pullthebox", email = "socrates.vlassis@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 4 - Beta",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: Software Development :: Embedded Systems :: Controller Area Network (CAN)",
    "Topic :: Software Development :: Embedded Systems :: Controller Area Network (CAN) :: J1939"
]
keywords = ["J1939", "CAN bus"]
requires-python = ">=3.13"
dependencies = [
    "dearpygui==2.0.0",
    "python-can==4.5.0",
    "openpyxl==3.1.5",
    "semver==3.0.4",
]

[project.scripts]
akrocansim = "akrocansim:main"

[project.urls]
Home = "https://github.com/cfsok/akrocansim"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/akrocansim/__init__.py"
